Farmer-aivision/requirements.txt
# ── Core ──────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1

# ── AI Inference ────────────────────────────────────────────────────
ultralytics>=8.1.0
numpy>=1.26.0
Pillow>=10.3.0

# ── Utilities ──────────────────────────────────────────────────────
aiofiles>=23.2.1
httpx>=0.27.0
tenacity>=8.2.3
